Rename files to the new insights ui convention - part 1
