+++ src/entry.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { BrowserRouter as Router } from 'react-router-dom';
 import { Provider } from 'react-redux';
-import store from './Utilities/store';
+import store from './utilities/store';
 import App from './App';
 
 const pathName = window.location.pathname.split('/');
