+++ src/smart-components/app-tabs/app-tabs.js
