+++ src/smart-components/app-tabs/apptabs.scss
