+++ src/smart-components/request/request.scss
@@ -1,5 +1,5 @@
 @import "../../App.scss";
-@import "../../PresentationalComponents/Shared/toolbarschema";
+@import "../../PresentationalComponents/shared/toolbarschema";
 
 .center_message {
   padding-top: 35px;
