+++ src/test/presentional-components/shared/404Route.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { mount } from 'enzyme';
 import toJson from 'enzyme-to-json';
-import NoMatch from '../../../PresentationalComponents/Shared/404Route';
+import NoMatch from '../../../PresentationalComponents/shared/404Route';
 
 describe('<NoMatch />', () => {
   it('should render correctly', () => {
