+++ src/test/presentional-components/shared/filter-toolbar-item.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { mount } from 'enzyme';
 import toJson from 'enzyme-to-json';
-import FilterToolbarItem from '../../../PresentationalComponents/Shared/FilterToolbarItem';
+import FilterToolbarItem from '../../../PresentationalComponents/shared/FilterToolbarItem';
 
 describe('<FilterToolbarItem />', () => {
   let intialProps;
